Revert to MV2 so script execution works
Manifest V3 borks this extension, so, in the way it is now, Chrome users will not be able to use jSin starting from some point in 2023, and as far as I can tell, Google is entirely to blame for it.
Since Manifest V2 will have completely lost support by then.

Firefox users *might* still be able to use it.